progress towards separate game and planarity packages
