+++ game/input_quit.py
@@ -1,5 +1,6 @@
 import sys, pygame
-import game_input
+
+from .ui_event import *
 
 
 def on_quit(event, game):
@@ -11,7 +12,7 @@ def on_key_down(event, game):
         sys.exit()
 
 
-game_input.subscribers.append({
+subscribers.append({
         pygame.QUIT: on_quit,
         pygame.KEYDOWN: on_key_down
     })
